separate parsing utils
move simple static functions
from _api to parsing_utils,
check that cleaned payload isn't empty
before creating attachment
